fix(core): print args warning when base is not passed
