+++ packages/workspace/src/command-line/utils.ts
@@ -2,6 +2,7 @@ import * as yargsParser from 'yargs-parser';
 import * as yargs from 'yargs';
 import * as fileUtils from '../core/file-utils';
 import { NxAffectedConfig } from '../core/shared-interfaces';
+import { output } from '../utils/output';
 
 const runOne = [
   'target',
@@ -102,6 +103,7 @@ export function splitArgsIntoNxArgsAndOverrides(
   }
 
   if (mode === 'affected') {
+    printArgsWarning(nxArgs);
     if (
       !nxArgs.files &&
       !nxArgs.uncommitted &&
@@ -139,5 +141,32 @@ export function getAffectedConfig(): NxAffectedConfig {
     return {
       defaultBase,
     };
+  }
+}
+
+function printArgsWarning(options: NxArgs) {
+  const { files, uncommitted, untracked, base, head, all } = options;
+  const affectedConfig = getAffectedConfig();
+
+  if (!files && !uncommitted && !untracked && !base && !head && !all) {
+    output.note({
+      title: `Affected criteria defaulted to --base=${output.bold(
+        `${affectedConfig.defaultBase}`
+      )} --head=${output.bold('HEAD')}`,
+    });
+  }
+
+  if (all) {
+    output.warn({
+      title: `Running affected:* commands with --all can result in very slow builds.`,
+      bodyLines: [
+        output.bold('--all') +
+          ' is not meant to be used for any sizable project or to be used in CI.',
+        '',
+        output.colors.gray(
+          'Learn more about checking only what is affected: '
+        ) + 'https://nx.dev/guides/monorepo-affected.',
+      ],
+    });
   }
 }
